Rework statystyk RJ
