+++ src/styles/global.scss
@@ -210,7 +210,7 @@ button {
 }
 
 button.btn--filled {
-  background-color: #333;
+  background-color: #1a1a1a;
   border-radius: 0.25em;
 
   &:hover {
